Unsuccessful code commit
I am trying to use the "frequency domain integration" method to calculate "velocity" and "displacement" based on "acceleration". Unfortunately, due to the influence of low-frequency noise caused by zero drift or other problems, it seems difficult to get correct results and use it for analysis. I hope to find a better way in the future.